chore(io): move _validate_schema_exists() from __csv_aux_writer to _pd_utils and use it in XML and CSV writers
