+++ src/pysdmx/io/_pd_utils.py
@@ -1,6 +1,7 @@
 import pandas as pd
 
 from pysdmx.errors import Invalid
+from pysdmx.io.pd import PandasDataset
 from pysdmx.model.concept import DataType
 from pysdmx.model.dataflow import Schema
 
@@ -79,3 +80,22 @@ def _validate_explicit_null_values(
                         f"Invalid null value 'NaN' in non-numeric component "
                         f"'{component.id}'."
                     )
+
+
+def _validate_schema_exists(dataset: PandasDataset) -> Schema:
+    """Validates that the dataset has a Schema defined.
+
+    Args:
+        dataset: The dataset to validate.
+
+    Returns:
+        The `Schema` from the dataset.
+
+    Raises:
+        Invalid: If the structure is not a `Schema`.
+    """
+    if not isinstance(dataset.structure, Schema):
+        raise Invalid(
+            "Dataset Structure is not a Schema. Cannot perform operation."
+        )
+    return dataset.structure
